requirements.txt
cryptography
feedparser==5.2.1
python-crontab
PyYAML